vet-edit-report - move _actions from appointments into vet/_lib/actions
