+++ src/app/vet/(logged-in)/appointments/_components/submit-report-form.tsx
@@ -31,7 +31,7 @@ import {
 import { zodResolver } from "@hookform/resolvers/zod";
 import { useForm } from "react-hook-form";
 import { z } from "zod";
-import { postBarkReportData } from "../_actions/post-bark-report-data";
+import { postBarkReportData } from "../../../_lib/actions/post-bark-report-data";
 import { useRouter } from "next/navigation";
 import { RoutePath } from "@/lib/route-path";
 import { capitalize } from "lodash";
